refactor: don't expose claimer registration in PUT interface
