+++ puffin/src/put.rs
@@ -83,14 +83,6 @@ pub trait Put<PB: ProtocolBehavior>:
 
     fn descriptor(&self) -> &AgentDescriptor;
 
-    /// Register a new claim for the agent
-    #[cfg(feature = "claims")]
-    fn register_claimer(&mut self);
-
-    /// Remove all claims in self
-    #[cfg(feature = "claims")]
-    fn deregister_claimer(&mut self);
-
     /// Returns a textual representation of the state in which self is
     fn describe_state(&self) -> &str;
