+++ tlspuffin/src/openssl/mod.rs
@@ -116,7 +116,6 @@ pub struct OpenSSL {
 
 impl Drop for OpenSSL {
     fn drop(&mut self) {
-        #[cfg(feature = "claims")]
         self.deregister_claimer();
     }
 }
@@ -157,7 +156,6 @@ impl Put<TLSProtocolBehavior> for OpenSSL {
     fn reset(&mut self, new_name: AgentName) -> Result<(), Error> {
         self.config.descriptor.name = new_name;
 
-        #[cfg(feature = "claims")]
         self.deregister_claimer();
 
         if self.config.use_clear {
@@ -168,14 +166,6 @@ impl Put<TLSProtocolBehavior> for OpenSSL {
             })?;
         }
 
-        // FIXME don't force-register a new claimer on reset
-        //
-        //    Because OpenSSL vendor libraries crash when no claimer is
-        //    registered (#253), we are forced to register a new one on
-        //    reset. This should be removed once this bug is fixed.
-        //
-        //    See the WolfSSL module for comparison.
-        #[cfg(feature = "claims")]
         self.register_claimer();
 
         Ok(())
@@ -183,42 +173,6 @@ impl Put<TLSProtocolBehavior> for OpenSSL {
 
     fn descriptor(&self) -> &AgentDescriptor {
         &self.config.descriptor
-    }
-
-    #[cfg(feature = "claims")]
-    fn register_claimer(&mut self) {
-        unsafe {
-            use foreign_types_openssl::ForeignTypeRef;
-
-            let agent_name = self.config.descriptor.name;
-            let claims = self.config.claims.clone();
-            let protocol_version = self.config.descriptor.tls_version;
-            let origin = self.config.descriptor.typ;
-
-            security_claims::register_claimer(
-                self.stream.ssl().as_ptr().cast(),
-                move |claim: security_claims::Claim| {
-                    if let Some(data) = claims_helpers::to_claim_data(protocol_version, claim) {
-                        claims
-                            .deref_borrow_mut()
-                            .claim_sized(crate::claims::TlsClaim {
-                                agent_name,
-                                origin,
-                                protocol_version,
-                                data,
-                            })
-                    }
-                },
-            );
-        }
-    }
-
-    #[cfg(feature = "claims")]
-    fn deregister_claimer(&mut self) {
-        unsafe {
-            use foreign_types_openssl::ForeignTypeRef;
-            security_claims::deregister_claimer(self.stream.ssl().as_ptr().cast());
-        }
     }
 
     fn describe_state(&self) -> &str {
@@ -276,7 +230,6 @@ impl OpenSSL {
             stream,
         };
 
-        #[cfg(feature = "claims")]
         openssl.register_claimer();
 
         Ok(openssl)
@@ -387,6 +340,42 @@ impl OpenSSL {
 
         Ok(ssl)
     }
+
+    fn register_claimer(&mut self) {
+        unsafe {
+            use foreign_types_openssl::ForeignTypeRef;
+
+            use crate::claims::claims_helpers;
+
+            let agent_name = self.config.descriptor.name;
+            let claims = self.config.claims.clone();
+            let protocol_version = self.config.descriptor.tls_version;
+            let origin = self.config.descriptor.typ;
+
+            security_claims::register_claimer(
+                self.stream.ssl().as_ptr().cast(),
+                move |claim: security_claims::Claim| {
+                    if let Some(data) = claims_helpers::to_claim_data(protocol_version, claim) {
+                        claims
+                            .deref_borrow_mut()
+                            .claim_sized(crate::claims::TlsClaim {
+                                agent_name,
+                                origin,
+                                protocol_version,
+                                data,
+                            })
+                    }
+                },
+            );
+        }
+    }
+
+    fn deregister_claimer(&mut self) {
+        unsafe {
+            use foreign_types_openssl::ForeignTypeRef;
+            security_claims::deregister_claimer(self.stream.ssl().as_ptr().cast());
+        }
+    }
 }
 
 pub enum MaybeError {
@@ -424,121 +413,6 @@ impl Into<Result<(), Error>> for MaybeError {
         match self {
             MaybeError::Ok => Ok(()),
             MaybeError::Err(err) => Err(err),
-        }
-    }
-}
-
-#[cfg(feature = "claims")]
-mod claims_helpers {
-    use puffin::agent::TLSVersion;
-    use smallvec::SmallVec;
-
-    use crate::claims::{
-        ClaimData, ClaimDataMessage, ClaimDataTranscript, Finished, TlsTranscript,
-        TranscriptCertificate, TranscriptClientFinished, TranscriptClientHello,
-        TranscriptPartialClientHello, TranscriptServerFinished, TranscriptServerHello,
-    };
-
-    pub fn to_claim_data(
-        protocol_version: TLSVersion,
-        claim: security_claims::Claim,
-    ) -> Option<ClaimData> {
-        match claim.typ {
-            // Transcripts
-            security_claims::ClaimType::CLAIM_TRANSCRIPT_CH => Some(ClaimData::Transcript(
-                ClaimDataTranscript::ClientHello(TranscriptClientHello(TlsTranscript(
-                    claim.transcript.data,
-                    claim.transcript.length,
-                ))),
-            )),
-            security_claims::ClaimType::CLAIM_TRANSCRIPT_PARTIAL_CH => Some(ClaimData::Transcript(
-                ClaimDataTranscript::PartialClientHello(TranscriptPartialClientHello(
-                    TlsTranscript(claim.transcript.data, claim.transcript.length),
-                )),
-            )),
-            security_claims::ClaimType::CLAIM_TRANSCRIPT_CH_SH => Some(ClaimData::Transcript(
-                ClaimDataTranscript::ServerHello(TranscriptServerHello(TlsTranscript(
-                    claim.transcript.data,
-                    claim.transcript.length,
-                ))),
-            )),
-            security_claims::ClaimType::CLAIM_TRANSCRIPT_CH_SERVER_FIN => {
-                Some(ClaimData::Transcript(ClaimDataTranscript::ServerFinished(
-                    TranscriptServerFinished(TlsTranscript(
-                        claim.transcript.data,
-                        claim.transcript.length,
-                    )),
-                )))
-            }
-            security_claims::ClaimType::CLAIM_TRANSCRIPT_CH_CLIENT_FIN => {
-                Some(ClaimData::Transcript(ClaimDataTranscript::ClientFinished(
-                    TranscriptClientFinished(TlsTranscript(
-                        claim.transcript.data,
-                        claim.transcript.length,
-                    )),
-                )))
-            }
-            security_claims::ClaimType::CLAIM_TRANSCRIPT_CH_CERT => Some(ClaimData::Transcript(
-                ClaimDataTranscript::Certificate(TranscriptCertificate(TlsTranscript(
-                    claim.transcript.data,
-                    claim.transcript.length,
-                ))),
-            )),
-            // Messages
-            // Transcripts in these messages are not up-to-date. They get updated after the Message
-            // has been processed
-            security_claims::ClaimType::CLAIM_FINISHED => {
-                Some(ClaimData::Message(ClaimDataMessage::Finished(Finished {
-                    outbound: claim.write > 0,
-                    client_random: SmallVec::from(claim.client_random.data),
-                    server_random: SmallVec::from(claim.server_random.data),
-                    session_id: SmallVec::from_slice(
-                        &claim.session_id.data[..claim.session_id.length as usize],
-                    ),
-                    authenticate_peer: false,             // FIXME
-                    peer_certificate: Default::default(), // FIXME
-                    master_secret: match protocol_version {
-                        TLSVersion::V1_3 => SmallVec::from_slice(&claim.master_secret.secret),
-                        TLSVersion::V1_2 => SmallVec::from_slice(&claim.master_secret_12.secret),
-                    },
-                    chosen_cipher: claim.chosen_cipher.data,
-                    available_ciphers: SmallVec::from_iter(
-                        claim.available_ciphers.ciphers[..claim.available_ciphers.length as usize]
-                            .iter()
-                            .map(|cipher| cipher.data),
-                    ),
-                    signature_algorithm: claim.signature_algorithm,
-                    peer_signature_algorithm: claim.peer_signature_algorithm,
-                })))
-            }
-            security_claims::ClaimType::CLAIM_CLIENT_HELLO => None,
-            security_claims::ClaimType::CLAIM_CCS => None,
-            security_claims::ClaimType::CLAIM_END_OF_EARLY_DATA => None,
-            security_claims::ClaimType::CLAIM_CERTIFICATE => None,
-            security_claims::ClaimType::CLAIM_KEY_EXCHANGE => None,
-            // FIXME it is weird that this returns the correct transcript
-            security_claims::ClaimType::CLAIM_CERTIFICATE_VERIFY => {
-                if claim.write == 0 {
-                    Some(ClaimData::Transcript(ClaimDataTranscript::ServerFinished(
-                        TranscriptServerFinished(TlsTranscript(
-                            claim.transcript.data,
-                            claim.transcript.length,
-                        )),
-                    )))
-                } else {
-                    None
-                }
-            }
-            security_claims::ClaimType::CLAIM_KEY_UPDATE => None,
-            security_claims::ClaimType::CLAIM_HELLO_REQUEST => None,
-            security_claims::ClaimType::CLAIM_SERVER_HELLO => None,
-            security_claims::ClaimType::CLAIM_CERTIFICATE_REQUEST => None,
-            security_claims::ClaimType::CLAIM_SERVER_DONE => None,
-            security_claims::ClaimType::CLAIM_SESSION_TICKET => None,
-            security_claims::ClaimType::CLAIM_CERTIFICATE_STATUS => None,
-            security_claims::ClaimType::CLAIM_EARLY_DATA => None,
-            security_claims::ClaimType::CLAIM_ENCRYPTED_EXTENSIONS => None,
-            _ => None,
         }
     }
 }
