+++ tlspuffin/src/tcp/mod.rs
@@ -308,16 +308,6 @@ impl Put<TLSProtocolBehavior> for TcpServerPut {
         &self.agent_descriptor
     }
 
-    #[cfg(feature = "claims")]
-    fn register_claimer(&mut self) {
-        panic!("Claims are not supported with TcpPut")
-    }
-
-    #[cfg(feature = "claims")]
-    fn deregister_claimer(&mut self) {
-        panic!("Claims are not supported with TcpPut")
-    }
-
     fn describe_state(&self) -> &str {
         panic!("Not supported")
     }
@@ -358,16 +348,6 @@ impl Put<TLSProtocolBehavior> for TcpClientPut {
 
     fn descriptor(&self) -> &AgentDescriptor {
         &self.agent_descriptor
-    }
-
-    #[cfg(feature = "claims")]
-    fn register_claimer(&mut self) {
-        panic!("Claims are not supported with TcpPut")
-    }
-
-    #[cfg(feature = "claims")]
-    fn deregister_claimer(&mut self) {
-        panic!("Claims are not supported with TcpPut")
     }
 
     fn describe_state(&self) -> &str {
